Use subcommands (split and ssupdate).
